Fixing relation on both sides

+++ Article/src/main/java/com/monkhub/article/controller/ArticleController.java
@@ -27,7 +27,7 @@ public class ArticleController {
 	}
 
 	@GetMapping("getArticleById/{articleId}")
-	public Article getArticleById(@PathVariable int articleId) {
+	public Article getArticleById(@PathVariable Long articleId) {
 		try {
 			return articleService.getArticleById(articleId);
 		} catch (NoSuchElementException e) {
