+++ Article/src/main/java/com/monkhub/article/service/ArticleService.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface ArticleService {
 
-	public boolean articleExistsById(int articleId);
+	public boolean articleExistsById(Long articleId);
 
 	public List<Article> getArticles();
 
-	public Article getArticleById(int articleId);
+	public Article getArticleById(Long articleId);
 
 	public void saveArticle(Article article);
